requirements.txt
coverage>=4.5.4
email-validator==1.1.1
Faker==4.1.0
flake8==3.7.9
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-paginate==0.6.0
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
psycopg2-binary==2.8.4
Werkzeug==0.16.1
